requirements.txt
aniso8601==8.1.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pptree==3.1
python-chess==1.999
pytz==2020.4
six==1.15.0
Werkzeug==1.0.1
